# perf: add benchee for measuring new controller actions