Adding bench-marking tool to be able to measure things, new controller actions in particular.

This is how the output would look like for one of them:

```
Operating System: Linux
CPU Information: AMD Ryzen 5 5600X 6-Core Processor
Number of Available Cores: 12
Available memory: 62.80 GB
Elixir 1.10.4
Erlang 22.3.4.21

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 1 s
parallel: 1
inputs: none specified
Estimated total run time: 16 s

Benchmarking active_oracles...
Benchmarking active_oracles_v2...

Name                        ips        average  deviation         median         99th %
active_oracles_v2         21.81       45.86 ms     ±5.76%       45.49 ms       54.16 ms
active_oracles            18.36       54.46 ms    ±16.48%       51.43 ms       77.47 ms

Comparison: 
active_oracles_v2         21.81
active_oracles            18.36 - 1.19x slower +8.61 ms

Memory usage statistics:

Name                      average  deviation         median         99th %
active_oracles_v2       499.80 KB     ±0.02%      499.86 KB      499.86 KB
active_oracles          561.06 KB     ±0.07%      560.99 KB      562.17 KB

Comparison: 
active_oracles_v2       499.86 KB
active_oracles          561.06 KB - 1.12x memory usage +61.26 KB

```